🏗 refactor(mcp): extract agentFetch and fetchWithRetry from shared.ts god module (#11530)
Split two cleanly-bounded concerns out of hooks/mcp/shared.ts (2085 lines):

- agentFetch.ts: agent token lifecycle + agentFetch() auth-injecting wrapper
  (~125 lines). shared.ts imports and re-exports for backward compat.

- fetchWithRetry.ts: FetchWithRetryOptions + fetchWithRetry() generic retry
  wrapper with exponential backoff (~100 lines). No shared.ts internals needed.

- clusterUtils.ts: re-export barrel for clusterDisplayName,
  shareMetricsBetweenSameServerClusters, deduplicateClustersByServer.
  Phase 2 will flip direction (implementations move here, shared.ts imports).

All 21+ existing imports from './mcp/shared' continue to work unchanged via
re-exports. No behavior changes — pure structural refactor.

Fixes #11530 (Phase 1)
