+++ web/src/hooks/mcp/clusterUtils.ts
@@ -1,0 +1,13 @@
+/**
+ * clusterUtils — convenience re-export barrel for cluster utility functions.
+ *
+ * Phase 1: Re-exports from shared.ts so consumers can import from a stable
+ *          path without pulling in the entire cluster cache.
+ * Phase 2 (#11530): Flip direction — move implementations here, make shared.ts
+ *                   import from this module.
+ */
+export {
+  clusterDisplayName,
+  shareMetricsBetweenSameServerClusters,
+  deduplicateClustersByServer,
+} from './shared'
